Homework Day1 Day2
I have not finished Day1 yet because I have not found the question which make me excited on that website. Maybe I need more time to survey the question.

The Day2's question is about my job and it is why I attend this activity. I want to survey a new way to help my team to finish the works.